+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TokenUpdateTransaction.java
@@ -2,15 +2,18 @@ package com.hedera.hashgraph.sdk;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.StringValue;
-import com.hedera.hashgraph.sdk.proto.*;
+import com.hedera.hashgraph.sdk.proto.SchedulableTransactionBody;
+import com.hedera.hashgraph.sdk.proto.TokenServiceGrpc;
+import com.hedera.hashgraph.sdk.proto.TokenUpdateTransactionBody;
+import com.hedera.hashgraph.sdk.proto.TransactionBody;
 import com.hedera.hashgraph.sdk.proto.TransactionResponse;
 import io.grpc.MethodDescriptor;
 import org.threeten.bp.Duration;
 import org.threeten.bp.Instant;
 
+import javax.annotation.Nullable;
 import java.util.LinkedHashMap;
 import java.util.Objects;
-import javax.annotation.Nullable;
 
 public class TokenUpdateTransaction extends Transaction<TokenUpdateTransaction> {
     @Nullable
@@ -227,6 +230,9 @@ public class TokenUpdateTransaction extends Transaction<TokenUpdateTransaction> 
 
     void initFromTransactionBody() {
         var body = sourceTransactionBody.getTokenUpdate();
+        if(body.hasToken()) {
+            tokenId = TokenId.fromProtobuf(body.getToken());
+        }
         if (body.hasTreasury()) {
             treasuryAccountId = AccountId.fromProtobuf(body.getTreasury());
         }