+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ContractCreateTransactionTest.java
@@ -9,6 +9,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class ContractCreateTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -42,5 +44,24 @@ public class ContractCreateTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytes() throws Exception {
+        var tx = new ContractCreateTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setBytecodeFileId(FileId.fromString("0.0.3003"))
+            .setAdminKey(unusedPrivateKey)
+            .setGas(0)
+            .setInitialBalance(Hbar.fromTinybars(1000))
+            .setProxyAccountId(AccountId.fromString("0.0.1001"))
+            .setAutoRenewPeriod(Duration.ofHours(10))
+            .setConstructorParameters(new byte[]{10, 11, 12, 13, 25})
+            .setMaxTransactionFee(Hbar.fromTinybars(100_000))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = ContractCreateTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }