+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ContractExecuteTransactionTest.java
@@ -9,6 +9,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class ContractExecuteTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -39,5 +41,21 @@ public class ContractExecuteTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytes() throws Exception {
+        var tx = new ContractExecuteTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setContractId(ContractId.fromString("0.0.5007"))
+            .setGas(10)
+            .setPayableAmount(Hbar.fromTinybars(1000))
+            .setFunctionParameters(ByteString.copyFrom(new byte[]{24, 43, 11}))
+            .setMaxTransactionFee(Hbar.fromTinybars(100_000))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = ContractExecuteTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }