+++ sdk/src/test/java/com/hedera/hashgraph/sdk/FileUpdateTransactionTest.java
@@ -8,6 +8,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class FileUpdateTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -38,5 +40,21 @@ public class FileUpdateTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytes() throws Exception {
+        var tx = new FileUpdateTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setFileId(FileId.fromString("0.0.6006"))
+            .setExpirationTime(Instant.ofEpochSecond(1554158728))
+            .setContents(new byte[]{1, 2, 3, 4, 5})
+            .setMaxTransactionFee(Hbar.fromTinybars(100_000))
+            .setKeys(unusedPrivateKey)
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = FileUpdateTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }