+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ScheduleCreateTransactionTest.java
@@ -8,6 +8,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class ScheduleCreateTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -40,5 +42,23 @@ public class ScheduleCreateTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytes() throws Exception {
+        var transferTransaction = new TransferTransaction()
+            .addHbarTransfer(AccountId.fromString("0.0.555"), new Hbar(-10))
+            .addHbarTransfer(AccountId.fromString("0.0.333"), new Hbar(10));
+        var tx = transferTransaction.schedule()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setAdminKey(unusedPrivateKey)
+            .setPayerAccountId(AccountId.fromString("0.0.222"))
+            .setScheduleMemo("hi")
+            .setMaxTransactionFee(new Hbar(1))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = ScheduleCreateTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }