+++ sdk/src/test/java/com/hedera/hashgraph/sdk/SystemDeleteTransactionTest.java
@@ -8,6 +8,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class SystemDeleteTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -50,5 +52,33 @@ public class SystemDeleteTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytesContract() throws Exception {
+        var tx = new SystemDeleteTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setContractId(ContractId.fromString("0.0.444"))
+            .setExpirationTime(validStart)
+            .setMaxTransactionFee(new Hbar(1))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = ScheduleDeleteTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
+    }
+
+    @Test
+    void shouldBytesFile() throws Exception {
+        var tx = new SystemDeleteTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setFileId(FileId.fromString("0.0.444"))
+            .setExpirationTime(validStart)
+            .setMaxTransactionFee(new Hbar(1))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = SystemDeleteTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }