+++ sdk/src/test/java/com/hedera/hashgraph/sdk/TokenBurnTransactionTest.java
@@ -8,6 +8,8 @@ import org.threeten.bp.Instant;
 
 import java.util.Collections;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class TokenBurnTransactionTest {
     private static final PrivateKey unusedPrivateKey = PrivateKey.fromString(
         "302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10");
@@ -50,5 +52,33 @@ public class TokenBurnTransactionTest {
             .sign(unusedPrivateKey)
             .toString()
         ).toMatchSnapshot();
+    }
+
+    @Test
+    void shouldBytesFungible() throws Exception {
+        var tx = new TokenBurnTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setTokenId(TokenId.fromString("0.0.111"))
+            .setAmount(30)
+            .setMaxTransactionFee(new Hbar(1))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = TokenBurnTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
+    }
+
+    @Test
+    void shouldBytesNft() throws Exception {
+        var tx = new TokenBurnTransaction()
+            .setNodeAccountIds(Collections.singletonList(AccountId.fromString("0.0.5005")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setTokenId(TokenId.fromString("0.0.111"))
+            .setSerials(Collections.singletonList(444L))
+            .setMaxTransactionFee(new Hbar(1))
+            .freeze()
+            .sign(unusedPrivateKey);
+        var tx2 = TokenBurnTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }